change the package name from schema to symbol
